+++ extension/src/panel/providers/index.ts
@@ -1,14 +1,11 @@
 export { ForkProvider } from './ForkProvider'
 export { isConnected } from './isConnected'
-export {
-  ProvideProvider,
-  useProvider,
-  useSubmitTransactions,
-} from './ProvideProvider'
+export { ProvideProvider, useProvider } from './ProvideProvider'
 export {
   getEip1193ReadOnlyProvider,
   getReadOnlyProvider,
 } from './readOnlyProvider'
+export { useSubmitTransactions } from './SubmitTransactionContext'
 export {
   ProvideInjectedWallet,
   useInjectedWallet,
